Changes to this commit renamed:    src/App.tsx -> src/components/App.tsx 	Moved into the components folder. modified:   src/components/HeroDetail.tsx 	Instead of rendering Animation component I used NotFound 	component instead. modified:   src/components/NotFound.tsx 	This component is now implemented the Animation component. modified:   src/index.tsx 	For other routes(*) render NotFound component.
